+++ handsontable/src/dataMap/dataMap.ts
@@ -595,7 +595,7 @@ class DataMap {
    * @param {number} [index] Visual index of the row to be removed. If not provided, the last row will be removed.
    * @param {number} [amount=1] Amount of the rows to be removed. If not provided, one row will be removed.
    * @param {string} [source] Source of method call.
-   * @returns {boolean} Returns `false` when action was cancelled, otherwise `true`.
+   * @returns {boolean} Returns `false` when action was canceled, otherwise `true`.
    */
   removeRow(index: number, amount = 1, source: string) {
     let rowIndex = Number.isInteger(index) ? index : -amount; // -amount = taking indexes from the end.
@@ -648,7 +648,7 @@ class DataMap {
    * @param {number} [index] Visual index of the column to be removed. If not provided, the last column will be removed.
    * @param {number} [amount=1] Amount of the columns to be removed. If not provided, one column will be removed.
    * @param {string} [source] Source of method call.
-   * @returns {boolean} Returns `false` when action was cancelled, otherwise `true`.
+   * @returns {boolean} Returns `false` when action was canceled, otherwise `true`.
    */
   removeCol(index: number, amount = 1, source: string) {
     if (this.hot!.dataType === 'object' || this.tableMeta.columns) {